+++ Css/Sections/Chatbot/chatbot.css
@@ -1,0 +1,98 @@
+/* Chatbot Styles*/
+.chatbot {
+    background-color:  var(--clr-default);
+    position: fixed;
+    z-index: 1;
+    right: 40px;
+    bottom: 100px;
+    width: 300px;
+    overflow: hidden;
+    border-radius: var(--brd-radius);
+    box-shadow:  0 0 128px 0 rgba(0, 0, 0, 0.1), 0 32px 64px -48px rgba(0, 0, 0, 0.5);
+}
+
+.chatbot header {
+    background-color: var(--clr-primary);
+    padding: 16px 0;
+    text-align: center;
+    border-top-left-radius: var(--brd-radius);
+    border-top-right-radius: var(--brd-radius);
+}
+
+.chatbot header h2 {
+    color: var(--clr-default);
+    font-size: 1.4rem;
+   
+}
+.chatbot .chatbox {
+    height: 310px !important;
+    overflow-y: auto;
+    padding:  30px 20px 70px;
+}
+.chatbox .chat {
+    display: flex;
+}
+.chatbox .incoming i {
+    height: 32px;
+    width: 32px;
+    background: var(--clr-primary);
+    color: var(--clr-default);
+    line-height:  32px;
+    text-align: center;
+    border-radius: var(--brd-radius);
+    align-self: flex-end;
+    margin: 0 10px 7px 0;
+}
+.chatbox .outgoing {
+    margin: 20px 0;
+    justify-content: flex-end;
+}
+.chatbox .chat p{
+    color: var(--clr-default);
+    font-size: 0.9rem;
+    max-width: 70%;
+    padding: 12px 16px;
+    border-radius:  10px 10px 0 10px;
+    background:  var(--clr-primary);
+}
+.chatbox .incoming p {
+    color: #000;
+    background:  #f2f2f2;
+    border-radius:  10px 10px 10 0;
+}
+.chat-input {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    display: flex;
+    gap: 5px;
+    background: var(--clr-primary);
+    border-top: 1px solid #ccc;
+}
+
+.chat-textrarea {
+    height: 55px;
+    width: 100%;
+    font-size: 0.95rem;
+    margin-left: 5px;
+    margin-bottom: 3px;
+    outline: none;
+    border: none;
+    resize: none;
+    background: var(--clr-primary);
+    color: var(--clr-default);
+    padding: 16px 15px 16px 0;
+}
+.send {
+    align-self: flex-end;
+    margin-bottom: 15px;
+    margin-right: 5px;
+    background: var(--clr-bg);
+    padding: 7px;
+    border-radius: var(--brd-radius);
+    color: var(--clr-default);
+    cursor: pointer;
+    
+    
+}
+/* End of Chatbot Style*/
